requirements.txt
anthropic>=0.40.0
alpaca-py>=0.30.0
ib_insync>=0.9.86
yfinance>=0.2.40
pandas>=2.0.0
pyarrow>=14.0.0
tqdm>=4.66.0
requests>=2.31.0
